Fix parent class conjunctions
A number of SBOL3 properties can appear on different classes. The
ontology should specify that these properties can appear on one OR the
other of the classes rather than requiring that they appear only on
objects that have ALL of the named classes. By default the domains are
conjunctions, specifying AND rather than OR. Explicitly use an OR on
the domains.